Rewrote property storage, now using a DBM
The old property caused a large I/O overhead, since property files are
unlinked and re-written as soon as a node with properties would be
touched in a revision. For repositories with many identical properties,
disk space has been wasted as well.

The new implementation identifies property data by its MD5-sum and
utilizes a DBM for storing the property data. Duplicate properties are
shared between nodes in order to save memory and disk space.