Added delayed deletion of property data
With the current design, the reference count of property data will
temporarily fall to zero if the respective node is touched during the
delta operation and holds the only reference. This change adds a garbage
collection operation after the delta editor has finished that will
remove unreferenced items from the property storage.