delta: Fix edge case producing a dump not loadable by svnadmin
When performing a dump not starting at initial revision and the first dump
revision contains a delete operation on a path added in an ancestor revision,
ensure to not insert a delete node in the dump file as no corresponding add
node was inserted for it.

Fixes https://github.com/jgehring/rsvndump/issues/18.